Move SlidingWindowAggregationStep to reader-jdbc
